formatter/requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
Markdown==3.4.1
python-dotenv==0.20.0
PyYAML==6.0